dmenu: reorganize includes and decls.
Add menu variable to Anchor class to remove one more global.
